data update occurs automatically without rebooting
from now on, the page is not rerendered,
a request will be sent to receive a new list of products
1. if the products are updated, a transparent block
and a progress bar appear on the site (body.active--load_data).
2. A new progress-bar module has been added to the material file.
3. The request method to the server has been changed from get to post
in the body of which the field indicates the need for relevant data.
If true, the server will send data for filters