update product data
Description: the "createProduct" function will send data to the server
without checking whether the data has been changed or not,
it only checks the validity of the data.
And the server will update only the changed data.

1. In the "new-photo" file, the container displays 3 photo formats
(link to a photo, user uploaded photo, 300x300 placeholder)
2. links to files can also be stored in an array of files >
(file, link, file, file)
3. the "imageSrc" variable stores links to photos of products
that the user has not changed