function createCharacteristics
Description: the function accepts two parameters:
product category and characteristics in string format.
It responds with an array of characteristics in the format of numbers
and an object in the format of the name of the characteristic
and the selected value, or an error message.

Else
1. Rename transformCharacteristicsStringToArray to createCharacteristics
2. "createCharacteristics" method has been added
to the "createProduct" and "updateProduct" functions